Extract signal parser, parameterize API status, tie status cap to window
Move the output parser out of the 1718-line agent-interface file. New leaf
modules signals.ts (the marker/error vocabulary) and output-signals.ts (the
AgentOutputSignals parser) break the import cycle. agent-interface re-exports
the moved symbols so existing importers are untouched.

Drop the redundant API_ERROR_401/429 table rows. API_ERROR is the one needle
and the status is a parameter: hasApiError() and hasApiErrorStatus(code).

Tie the status FIFO cap to the render window. New constants.ts holds
COLLAPSED_COUNT/EXPANDED_COUNT; the store caps at EXPANDED_COUNT instead of an
arbitrary 100, since the status bar is the only consumer and shows at most that
many lines.

Generated-By: PostHog Code
Task-Id: dc10b8c8-7d3c-494e-a7e2-95a9ca412017
